CustomVision/requirements.txt
azure-storage-blob
PIL
requests
msrest
azure-cognitiveservices-vision-customvision